+++ backend/src/main/java/com/wooteco/nolto/auth/ui/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.wooteco.nolto.user.ui.dto;
+package com.wooteco.nolto.auth.ui.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
